Fix and slipt tests, remove as any.
